pyproject.toml
[project]
name = "difonlib"
version = "0.2.2"
description = "python libraries"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "evdev>=1.9.2",
    "nicegui>=3.2.0",
    "pexpect>=4.9.0",
    "pyyaml>=6.0.3",
    "tinytuya>=1.17.4",
    "types-pexpect>=4.9.0.20250916",
    "types-PyYAML>=6.0.12.20250915",
    "types-xmltodict>=1.0.1.20250920",
    "xmltodict>=1.0.2",
]

[tool.uv]
# Здесь uv хранит настройки окружения, зависимостей и сборки
package = true

[tool.mypy]
# 🧩 Проверяем только свой код
disable_error_code = ["import-untyped"]
ignore_missing_imports = true
follow_imports = "silent"
no_site_packages = true

# 🚫 Строгие правила для собственного кода
disallow_untyped_defs = true
warn_unused_ignores = true
warn_return_any = true

# 📦 Если проект лежит в src/
mypy_path = ["src"]


[tool.pytest.ini_options]
addopts = "-v"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "ruff>=0.6.0",
    "black>=24.8.0",
    "mypy>=1.11.0",
    "twine>=5.1.1",
]

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100

